More insert.

+++ FILE_Synchronize.py
@@ -9,6 +9,11 @@ hostname = os.environ["HOST"]
 if hostname == "":
     hostname = os.environ["HOSTNAME"]
 
+
+if hostname == "":
+    print("We have found |hostname|=", len(hostname), " and hostname=", hostname)
+    os.exit(0)
+
 print("Finding the characteristic entry to be hostname=", hostname)
 
 
@@ -28,6 +33,10 @@ TripleInfo = namedtuple('TripleInfo', 'filename date hashv')
 # triple: filename, hash, date
 
 
+#
+# Main basic routines of the system
+#
+
 def get_hash(filename):
     h = hashlib.sha1()
     with open(filename,'rb') as file:
@@ -39,8 +48,31 @@ def get_hash(filename):
    return h.hexdigest()
 
 
+def wait_for_exist_file(list_file):
+    while True:
+        evalue = True
+        for e_file in list_file:
+            if not os.path.exists(e_file):
+                evalue = False
+        if evalue:
+            return
+        time.sleep(2)
 
-def send_file(e_triple, prev_triple):
+def wait_for_non_exist_file(list_file):
+    while True:
+        evalue = True
+        for e_file in list_file:
+            if os.path.exists(e_file):
+                evalue = False
+        if evalue:
+            return
+        time.sleep(2)
+
+#
+# Atomic operation.
+#
+
+def send_request(do_rename, prev_triple, e_triple):
     FileCopy1 = DropboxAddress + "FILEO_Meta_" + hostname
     FileCopy2 = DropboxAddress + "FILEO_data_" + hostname
     while True:
@@ -50,6 +82,10 @@ def send_file(e_triple, prev_triple):
     print("Synchronization files missing. We can proceed")
     #
     os.open(FileCopy1, 'w')
+    if do_rename:
+        os.write("1\n")
+    else:
+        os.write("2\n")
     os.write(prev_triple.filename) # previous file
     os.write(prev_triple.date) # previous hash
     os.write(prev_triple.hashv) # previous date
@@ -59,12 +95,14 @@ def send_file(e_triple, prev_triple):
     os.write(e_triple.hashv) # previous date
     os.close()
     #
-    os.system("cp " + DirectorySync + e_file, " " + FileCopy2)
+    if not do_rename:
+        os.system("cp " + DirectorySync + e_file, " " + FileCopy2)
 
-def recv_file(remote_hostname):
-    FileCopy1 = DropboxAddress + "FILEO_Meta_" + remotehostname
-    FileCopy2 = DropboxAddress + "FILEO_data_" + remotehostname
+def recv_request(remote_hostname):
+    FileCopy1 = DropboxAddress + "FILEO_Meta_" + remote_hostname
+    FileCopy2 = DropboxAddress + "FILEO_data_" + remote_hostname
     os.open(FileCopy1, 'r')
+    evalue = os.read()
     prev_file = os.read()
     prev_hash = os.read()
     prev_date = os.read()
@@ -74,20 +112,23 @@ def recv_file(remote_hostname):
     e_date = os.read()
     os.close()
     #
-    if prev_file in MapFile.keys():
-        last_ent = MapFile[prev_file][-1]
-        if last_ent[0] == prev_hash and last_ent[1] == prev_date:
-            os.system("mv " + FileCopy2 + " " + prevFile)
-            MapFile[prev_file].append([e_hash, e_date])
+    if evalue == 2:
+        if prev_file in MapFile.keys():
+            last_ent = MapFile[prev_file][-1]
+            if last_ent[0] == prev_hash and last_ent[1] == prev_date:
+                os.system("mv " + FileCopy2 + " " + prevFile)
+                MapFile[prev_file].append([e_hash, e_date])
+            else:
+                print("Incoherent changes of the file database")
+                sys.exit(0);
         else:
-            print("Incoherent changes of the file database")
-            sys.exit(0);
+            os.system("mv " + FileCopy2 + " " + prevFile)
+            MapFile[prev_file] = [ [e_hash, e_date] ]
+            #
+        os.remove(FileCopy1)
+        os.remove(FileCopy2)
     else:
-        os.system("mv " + FileCopy2 + " " + prevFile)
-        MapFile[prev_file] = [ [e_hash, e_date] ]
-    #
-    os.remove(FileCopy1)
-    os.remove(FileCopy2)
+        
 
 
 def read_single_file(e_file):
@@ -98,11 +139,15 @@ def read_single_file(e_file):
         MapFiles[e_file].append([e_hash, e_date] )
     else:
         MapFiles[e_file] = [  [e_hash, e_date]  ]
+    e_triple = TripleInfo(e_file, e_date, e_hash)
 
 
 def read_all_files(e_dir):
-    list_files = sys.list_all_files_and_subdirectory(e_dir)
-    for e_file in list_files:
+    list_infos = []
+    for root, subdirs, files in os.walk(e_dir):
+        list_infos.append( [root, subdirs, files])
+    for e_info in list_infos:
+        e_file = e_info[2]
         read_single_file(e_file)
 
 while True:
